Building two versions of Data Access class
In this app one version will be for stub testing while the other will
eventually connect to a database.
Set the default and alternative version to be injected into the service
class so just one version is actually injected.
Created the META-INF folder with beans.xml configuration file for the
alternative class and included it in the build.xml
The testing version doesn't have to be an EJB
Qualifier annotations can be used too, but I could not get them to work